Added a floor, hit's hitbox. changed the logic to be a switch case so I can handle the floor hitbox.
Next I'm going to work on making it so the user can't press and hold up, that it will only be a tick up on key down making the user have to tap the button like in the original.

Also, I'm going to make it so the player can't fly up infinantly. that it will only go as far as the top of the canvas an stop.

At that point I think it will be ready to push to the site.